+++ src/styles/chat-images.scss
@@ -1,8 +1,13 @@
 @import "spinner";
 @import "chat-message";
 
-#chat-controls .control-buttons {
-  flex: 0 0 72px;
+#chat-controls {
+  & .control-buttons {
+    flex: 0 0 72px;
+  }
+  & .ci-control-buttons {
+    flex: 0 0 24px;
+  }
 }
 
 #ci-upload-image-hidden-input {
